requirements.txt
# Core Flask dependencies
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.7
Flask-CORS==5.0.0
Flask-Mail==0.10.0

# Database
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
pymysql==1.1.1

# Authentication & Security
Werkzeug==3.1.3
PyJWT==2.10.1
python-dotenv==1.0.1
bcrypt==4.2.1
cryptography==43.0.3

# Forms and validation
WTForms==3.2.1
email-validator==2.2.0

# API and serialization
marshmallow==3.23.1
marshmallow-sqlalchemy==1.1.0

# Date and time
python-dateutil==2.9.0.post0

# Utilities
requests==2.32.3
Pillow==11.0.0

# Xero Integration
xero-python==5.0.0
oauthlib==3.2.2
requests-oauthlib==1.3.1

# Development tools
python-decouple==3.8
pytest==8.3.2

# Production server
gunicorn==23.0.0